#333 Seperated plan into seperate functions, beginning system tests
