+++ src/artemis/devices/oav/oav_detector.py
@@ -7,14 +7,16 @@ from ophyd import (
     AreaDetector,
     CamBase,
     Component,
+    Device,
     EpicsSignal,
-    EpicsSignalRO,
     HDF5Plugin,
     OverlayPlugin,
     ProcessPlugin,
     ROIPlugin,
 )
 
+from artemis.devices.backlight import Backlight
+from artemis.devices.motors import I03Smargon
 from artemis.devices.oav.grid_overlay import SnapshotWithGrid
 
 
@@ -33,8 +35,13 @@ class ColorMode(IntEnum):
     YUV421 = 7
 
 
-class Camera(CamBase):
-    zoom: EpicsSignal = Component(EpicsSignal, "FZOOM:ZOOMPOSCMD")
+class ZoomController(Device):
+    """
+    Device to control the zoom level, this is unfortunately on a different prefix
+    from CAM.
+    """
+
+    zoom: EpicsSignal = Component(EpicsSignal, "ZOOMPOSCMD")
 
 
 class EdgeOutputArrayImageType(IntEnum):
@@ -49,83 +56,42 @@ class EdgeOutputArrayImageType(IntEnum):
     CLOSED_EDGES = 4
 
 
-class OAV(AreaDetector):
-    # signal that was here before
-    # on: EpicsSignalRO = Component(EpicsSignalRO, "ProcessConnected_RBV")
+class MXSC(Device):
+    """
+    Device for edge detection plugin.
+    """
 
-    # snapshot PVs
-    cam: ADC = ADC(CamBase, "CAM:")
-    roi: ADC = ADC(ROIPlugin, "ROI:")
-    proc: ADC = ADC(ProcessPlugin, "PROC:")
-    over: ADC = ADC(OverlayPlugin, "OVER:")
-    tiff: ADC = ADC(OverlayPlugin, "TIFF:")
-    hdf5: ADC = ADC(HDF5Plugin, "HDF5:")
-    snapshot: SnapshotWithGrid = Component(SnapshotWithGrid, "MJPG")
-
-    # Edge detection PVs
-    colour_mode_pv: EpicsSignal = Component(EpicsSignal, "CAM:ColorMode")
-    x_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "MJPG:ArraySize1_RBV")
-    y_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "MJPG:ArraySize2_RBV")
-    input_rbpv: EpicsSignalRO = Component(EpicsSignalRO, "MJPG:NDArrayPort_RBV")
-    exposure_rbpv: EpicsSignalRO = Component(EpicsSignalRO, "CAM:AcquireTime_RBV")
-    acquire_period_rbpv: EpicsSignalRO = Component(
-        EpicsSignalRO, "CAM:AcquirePeriod_RBV"
+    input_plugin_pv: EpicsSignal = Component(EpicsSignal, "NDArrayPort")
+    enable_callbacks_pv: EpicsSignal = Component(EpicsSignal, "EnableCallbacks")
+    min_callback_time_pv: EpicsSignal = Component(EpicsSignal, "MinCallbackTime")
+    blocking_callbacks_pv: EpicsSignal = Component(EpicsSignal, "BlockingCallbacks")
+    read_file: EpicsSignal = Component(EpicsSignal, "ReadFile")
+    py_filename: EpicsSignal = Component(EpicsSignal, "Filename", string=True)
+    preprocess_operation: EpicsSignal = Component(EpicsSignal, "Preprocess")
+    preprocess_ksize: EpicsSignal = Component(EpicsSignal, "PpParam1")
+    canny_upper_threshold: EpicsSignal = Component(EpicsSignal, "CannyUpper")
+    canny_lower_threshold: EpicsSignal = Component(EpicsSignal, "CannyLower")
+    close_ksize: EpicsSignal = Component(EpicsSignal, "CloseKsize")
+    sample_detection_scan_direction: EpicsSignal = Component(
+        EpicsSignal, "ScanDirection"
     )
-    gain_rbpv: EpicsSignalRO = Component(EpicsSignalRO, "CAM:Gain_RBV")
-    input_pv: EpicsSignal = Component(EpicsSignal, "MJPG:NDArrayPort")
-    exposure_pv: EpicsSignal = Component(EpicsSignal, "CAM:AcquireTime")
-    acquire_period_pv: EpicsSignal = Component(EpicsSignal, "CAM:AcquirePeriod")
-    gain_pv: EpicsSignal = Component(EpicsSignal, "CAM:Gain")
-    enable_overlay_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:EnableCallbacks")
-    overlay_port_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:NDArrayPort")
-    use_overlay1_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:1:Use")
-    use_overlay2_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:Use")
-    overlay2_shape_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:Shape")
-    overlay2_red_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:Red")
-    overlay2_green_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:Green")
-    overlay2_blue_pv: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:Blue")
-    overlay2_x_position: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:PositionX")
-    overlay2_y_position: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:PositionY")
-    overlay2_x_size: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:SizeX")
-    overlay2_y_size: EpicsSignalRO = Component(EpicsSignalRO, "OVER:2:SizeY")
-
-    # MXSC signals
-    input_plugin_pv: EpicsSignal = Component(EpicsSignal, "MXSC:NDArrayPort")
-    enable_callbacks_pv: EpicsSignal = Component(EpicsSignal, "MXSC:EnableCallbacks")
-    min_callback_time_pv: EpicsSignal = Component(EpicsSignal, "MXSC:MinCallbackTime")
-    blocking_callbacks_pv: EpicsSignal = Component(
-        EpicsSignal, "MXSC:BlockingCallbacks"
+    sample_detection_min_tip_height: EpicsSignal = Component(
+        EpicsSignal, "MinTipHeight"
     )
-    read_file_pv: EpicsSignal = Component(EpicsSignal, "MXSC:ReadFile")
-    py_filename_pv: EpicsSignal = Component(EpicsSignal, "MXSC:Filename", string=True)
-    py_filename_rbpv: EpicsSignal = Component(
-        EpicsSignal, "MXSC:Filename_RBV", string=True
-    )
-    preprocess_operation_pv: EpicsSignal = Component(EpicsSignal, "MXSC:Preprocess")
-    preprocess_ksize_pv: EpicsSignal = Component(EpicsSignal, "MXSC:PpParam1")
-    canny_upper_threshold_pv: EpicsSignal = Component(EpicsSignal, "MXSC:CannyUpper")
-    canny_lower_threshold_pv: EpicsSignal = Component(EpicsSignal, "MXSC:CannyLower")
-    close_ksize_pv: EpicsSignal = Component(EpicsSignal, "MXSC:CloseKsize")
-    sample_detection_scan_direction_pv: EpicsSignal = Component(
-        EpicsSignal, "MXSC:ScanDirection"
-    )
-    sample_detection_min_tip_height_pv: EpicsSignal = Component(
-        EpicsSignal, "MXSC:MinTipHeight"
-    )
-    tip_x_pv: EpicsSignal = Component(EpicsSignal, "MXSC:TipX")
-    tip_y_pv: EpicsSignal = Component(EpicsSignal, "MXSC:TipY")
-    top_pv: EpicsSignal = Component(EpicsSignal, "MXSC:Top")
-    bottom_pv: EpicsSignal = Component(EpicsSignal, "MXSC:Bottom")
-    output_array_pv: EpicsSignal = Component(EpicsSignal, "MXSC:OutputArray")
-    draw_tip_pv: EpicsSignal = Component(EpicsSignal, "MXSC:DrawTip")
-    draw_edges_pv: EpicsSignal = Component(EpicsSignal, "MXSC:DrawEdges")
+    tip_x: EpicsSignal = Component(EpicsSignal, "TipX")
+    tip_y: EpicsSignal = Component(EpicsSignal, "TipY")
+    top: EpicsSignal = Component(EpicsSignal, "Top")
+    bottom: EpicsSignal = Component(EpicsSignal, "Bottom")
+    output_array: EpicsSignal = Component(EpicsSignal, "OutputArray")
+    draw_tip: EpicsSignal = Component(EpicsSignal, "DrawTip")
+    draw_edges: EpicsSignal = Component(EpicsSignal, "DrawEdges")
 
     def get_edge_waveforms(self):
         """
         Get the waveforms from the PVs as numpy arrays.
         """
-        yield from bps.rd(self.top_pv)
-        yield from bps.rd(self.bottom_pv)
+        yield from bps.rd(self.top)
+        yield from bps.rd(self.bottom)
 
     def get_edge_waveforms_as_numpy_arrays(self):
         return (np.array(pv) for pv in tuple(self.get_edge_waveforms()))
@@ -152,22 +118,33 @@ class OAV(AreaDetector):
         yield from bps.abs_set(self.blocking_callbacks_pv, 0)
 
         # Set the python file to use for calculating the edge waveforms
-        yield from bps.abs_set(self.py_filename_pv, filename, wait=True)
-        yield from bps.abs_set(self.read_file_pv, 1)
+        yield from bps.abs_set(self.py_filename, filename, wait=True)
+        yield from bps.abs_set(self.read_file, 1)
 
         # Image annotations
-        yield from bps.abs_set(self.draw_tip_pv, True)
-        yield from bps.abs_set(self.draw_edges_pv, True)
+        yield from bps.abs_set(self.draw_tip, True)
+        yield from bps.abs_set(self.draw_edges, True)
 
         # Use the original image type for the edge output array
-        yield from bps.abs_set(self.output_array_pv, EdgeOutputArrayImageType.ORIGINAL)
+        yield from bps.abs_set(self.output_array, EdgeOutputArrayImageType.ORIGINAL)
 
-    def get_sizes_from_pvs(self):
-        """
-        Yields the sizes from PVs.
-        Args: None
-        Yields:
-            The values from x_size_pv, y_size_pv.
-        """
-        yield from bps.rd(self.x_size_pv)
-        yield from bps.rd(self.y_size_pv)
+
+class OAV(AreaDetector):
+    cam: CamBase = ADC(CamBase, "-EA-OAV-01:CAM:")
+    roi: ADC = ADC(ROIPlugin, "-DI-OAV-01:ROI:")
+    proc: ADC = ADC(ProcessPlugin, "-DI-OAV-01:PROC:")
+    over: ADC = ADC(OverlayPlugin, "-DI-OAV-01:OVER:")
+    tiff: ADC = ADC(OverlayPlugin, "-DI-OAV-01:TIFF:")
+    hdf5: ADC = ADC(HDF5Plugin, "-DI-OAV-01:HDF5:")
+    snapshot: SnapshotWithGrid = Component(SnapshotWithGrid, "-DI-OAV-01:MJPG:")
+    mxsc: MXSC = ADC(MXSC, "-DI-OAV-01:MXSC:")
+    zoom_controller: ZoomController = ADC(ZoomController, "-EA-OAV-01-FZOOM:")
+
+
+if __name__ == "__main__":
+
+    beamline = "S03SIM"
+    smargon: I03Smargon = Component(I03Smargon, "-MO-SGON-01:")
+    backlight: Backlight = Component(Backlight, "-EA-BL-01:")
+    oav = OAV(name="oav", prefix=beamline)
+    oav.wait_for_connection()
