+++ src/artemis/devices/oav/snapshot.py
@@ -1,15 +1,20 @@
 import threading
 from pathlib import Path
 
+import bluesky.plan_stubs as bps
 import requests
-from ophyd import Component, Device, DeviceStatus, EpicsSignal, Signal
+from ophyd import Component, Device, DeviceStatus, EpicsSignal, EpicsSignalRO, Signal
 from PIL import Image
 
 
 class Snapshot(Device):
     filename: Signal = Component(Signal)
     directory: Signal = Component(Signal)
-    url: EpicsSignal = Component(EpicsSignal, ":JPG_URL_RBV", string=True)
+    url: EpicsSignal = Component(EpicsSignal, "JPG_URL_RBV", string=True)
+    x_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize1_RBV")
+    y_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize2_RBV")
+    input_rbpv: EpicsSignalRO = Component(EpicsSignalRO, "NDArrayPort_RBV")
+    input_pv: EpicsSignal = Component(EpicsSignal, "NDArrayPort")
     KICKOFF_TIMEOUT: float = 10.0
 
     def trigger(self):
@@ -36,3 +41,8 @@ class Snapshot(Device):
 
     def post_processing(self, image: Image.Image):
         pass
+
+    def get_sizes_from_pvs(self):
+        x_size = yield from bps.rd(self.x_size_pv)
+        y_size = yield from bps.rd(self.y_size_pv)
+        return x_size, y_size
